add @game.h
