+++ NetWorld-Immersion/NetWorld-Immersion.vcxproj
@@ -163,8 +163,11 @@
     <Image Include="..\Debug\UpperArmor__inventory_butt.png" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="BattleGame.h" />
     <ClInclude Include="bmp.h" />
     <ClInclude Include="enemy.h" />
+    <ClInclude Include="TerminalGame.h" />
+    <ClInclude Include="MapGame.h" />
     <ClInclude Include="mouse.h" />
     <ClInclude Include="player.h" />
     <ClInclude Include="timer.h" />
